+++ google/src/main/scala/org/broadinstitute/dsde/workbench/metrics/GoogleInstrumentedService.scala
@@ -5,7 +5,7 @@ package org.broadinstitute.dsde.workbench.metrics
   */
 object GoogleInstrumentedService extends Enumeration {
   type GoogleInstrumentedService = Value
-  val Billing, Storage, Genomics, Groups, PubSub, Dataproc = Value
+  val Billing, Storage, Genomics, Groups, PubSub, Dataproc, Iam = Value
 
   /**
     * Expansion for GoogleInstrumentedService which uses the default toString implementation.